introduce RepositoryFactory and builders; refactor packages
The ConfigurationRepository has to be a Spring Bean. However, as it is created by the user of the library, it is currently not possible for the repository to depend on Baigan-internal Spring beans without exposing those to the user.

  This PR introduces a `RepositoryFactory` as the only means to create repositories. This way, the repository factory can inject any required beans transparently, while the user configures the repository through the returned builder.

Furthermore, the package structure is refactored to streamline it.
