+++ src/main/java/org/zalando/baigan/context/ConfigurationContextProviderRegistryImpl.java
@@ -1,12 +1,12 @@
 /**
  * Copyright (C) 2015 Zalando SE (http://tech.zalando.com)
- *
+ * <p>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * <p>
  *         http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -16,20 +16,16 @@
 
 package org.zalando.baigan.context;
 
-import java.util.Collection;
-import java.util.Set;
-import java.util.function.Consumer;
-
-import javax.annotation.Nonnull;
-
-import org.springframework.stereotype.Component;
-import org.springframework.util.StringUtils;
-import org.zalando.baigan.provider.ContextProvider;
-
 import com.google.common.base.Preconditions;
 import com.google.common.collect.HashMultimap;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Multimap;
+import org.springframework.stereotype.Component;
+import org.springframework.util.StringUtils;
+
+import javax.annotation.Nonnull;
+import java.util.Collection;
+import java.util.Set;
 
 /**
  * Default implementation for {@link ContextProviderRegistry} and
@@ -49,13 +45,7 @@ public class ConfigurationContextProviderRegistryImpl
     public void register(@Nonnull final ContextProvider contextProvider) {
         Preconditions.checkNotNull(contextProvider,
                 "Attempt to register null as a context provider!");
-        contextProvider.getProvidedContexts().forEach(new Consumer<String>() {
-            @Override
-            public void accept(String context) {
-                providerMap.put(context, contextProvider);
-            }
-        });
-
+        contextProvider.getProvidedContexts().forEach(context -> providerMap.put(context, contextProvider));
     }
 
     @Override
