+++ src/main/java/org/zalando/baigan/context/ContextProvider.java
@@ -1,12 +1,12 @@
 /**
  * Copyright (C) 2015 Zalando SE (http://tech.zalando.com)
- *
+ * <p>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * <p>
  *         http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -14,28 +14,21 @@
  * limitations under the License.
  */
 
-package org.zalando.baigan.etcd.model;
+package org.zalando.baigan.context;
+
+import java.util.Set;
+
+import javax.annotation.Nonnull;
 
 /**
+ * Provides the value of the named context parameter at runtime.
+ *
  * @author mchand
+ *
  */
-public class DirResultNode {
-    private String action;
-    private DirNode node;
+public interface ContextProvider {
 
-    public String getAction() {
-        return action;
-    }
+    String getContextParam(@Nonnull final String name);
 
-    public void setAction(final String action) {
-        this.action = action;
-    }
-
-    public DirNode getNode() {
-        return node;
-    }
-
-    public void setNode(final DirNode node) {
-        this.node = node;
-    }
+    Set<String> getProvidedContexts();
 }
