+++ src/main/java/org/zalando/baigan/repository/S3ConfigurationRepository.java
@@ -1,15 +1,12 @@
-package org.zalando.baigan.service.aws;
+package org.zalando.baigan.repository;
 
 import com.amazonaws.services.kms.AWSKMS;
-import com.amazonaws.services.kms.AWSKMSClientBuilder;
 import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 import com.google.common.collect.ImmutableMap;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.zalando.baigan.model.Configuration;
-import org.zalando.baigan.service.AbstractConfigurationRepository;
-import org.zalando.baigan.service.ConfigurationRepository;
+import org.zalando.baigan.repository.aws.S3FileLoader;
 
 import javax.annotation.Nonnull;
 import java.util.List;
@@ -21,67 +18,18 @@ import java.util.concurrent.TimeUnit;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
 
+/**
+ * A {@link ConfigurationRepository} implementation that loads the configuration from an S3 bucket in regular
+ * intervals. It can read KMS-encrypted configuration files.
+ */
 public class S3ConfigurationRepository extends AbstractConfigurationRepository {
 
     private static final Logger LOG = LoggerFactory.getLogger(S3ConfigurationRepository.class);
-
-    /**
-     * The default refresh interval which is 60 seconds
-     */
-    private static final long DEFAULT_REFRESH_INTERVAL = 60;
 
     private final S3FileLoader s3Loader;
     private final long refreshInterval;
     private final ScheduledThreadPoolExecutor executor;
     private volatile Map<String, Configuration> configurationsMap = ImmutableMap.of();
-
-    /**
-     * @param bucketName The name of the bucket
-     * @param key        The object key, usually, the "full path" to the JSON file stored in the bucket
-     *
-     * @deprecated Use {@link S3ConfigurationRepositoryBuilder instead}
-     * <p>
-     * Provides a {@link ConfigurationRepository} that reads configurations from a JSON file stored in a S3 bucket.
-     * It refreshes configurations using the default refresh interval (60 seconds)
-     */
-    @Deprecated
-    public S3ConfigurationRepository(@Nonnull final String bucketName, @Nonnull final String key) {
-        this(bucketName, key, DEFAULT_REFRESH_INTERVAL);
-    }
-
-    /**
-     * @param bucketName      The name of the bucket
-     * @param key             The object key, usually, the "full path" to the JSON file stored in the bucket
-     * @param refreshInterval The interval, in seconds, to refresh the configurations. A value of 0 disables refreshing
-     *                        <p>
-     * @see #S3ConfigurationRepository(String, String)
-     *
-     * @deprecated Use {@link S3ConfigurationRepositoryBuilder instead}
-     * <p>
-     * Provides a {@link ConfigurationRepository} that reads configurations from a JSON file stored in a S3 bucket.
-     */
-    @Deprecated
-    public S3ConfigurationRepository(@Nonnull final String bucketName, @Nonnull final String key, final long refreshInterval) {
-        this(bucketName, key, refreshInterval, new ScheduledThreadPoolExecutor(1));
-    }
-
-    /**
-     * @param bucketName      The name of the bucket
-     * @param key             The object key, usually, the "full path" to the JSON file stored in the bucket
-     * @param refreshInterval The interval, in seconds, to refresh the configurations. A value of 0 disables refreshing
-     * @param executor        The executor to refresh the configurations in the specified interval
-     *                        <p>
-     * @see #S3ConfigurationRepository(String, String)
-     *
-     * @deprecated Use {@link S3ConfigurationRepositoryBuilder instead}
-     * <p>
-     * Provides a {@link ConfigurationRepository} that reads configurations from a JSON file stored in a S3 bucket.
-     */
-    @Deprecated
-    public S3ConfigurationRepository(@Nonnull final String bucketName, @Nonnull final String key,
-                                     final long refreshInterval, final ScheduledThreadPoolExecutor executor) {
-        this(bucketName, key, refreshInterval, executor, AmazonS3ClientBuilder.defaultClient(), AWSKMSClientBuilder.defaultClient());
-    }
 
     S3ConfigurationRepository(@Nonnull final String bucketName, @Nonnull final String key,
                               final long refreshInterval, final ScheduledThreadPoolExecutor executor,
