+++ src/main/java/org/zalando/baigan/repository/etcd/model/KeyResultNode.java
@@ -1,12 +1,12 @@
 /**
  * Copyright (C) 2015 Zalando SE (http://tech.zalando.com)
- *
+ * <p>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * <p>
  *         http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -14,19 +14,28 @@
  * limitations under the License.
  */
 
-package org.zalando.baigan.etcd.model;
+package org.zalando.baigan.repository.etcd.model;
 
 /**
  * @author mchand
  */
-public class KeyNode extends AbstractNode {
-    private String value;
+public class KeyResultNode {
+    private String action;
+    private KeyNode node;
 
-    public String getValue() {
-        return value;
+    public String getAction() {
+        return action;
     }
 
-    public void setValue(final String value) {
-        this.value = value;
+    public void setAction(final String action) {
+        this.action = action;
+    }
+
+    public KeyNode getNode() {
+        return node;
+    }
+
+    public void setNode(final KeyNode node) {
+        this.node = node;
     }
 }
