+++ src/test/java/org/zalando/baigan/proxy/ConfigurationServiceBeanFactoryIT.java
@@ -14,7 +14,7 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.zalando.baigan.BaiganSpringContext;
 import org.zalando.baigan.annotation.ConfigurationServiceScan;
 import org.zalando.baigan.fixture.SomeConfiguration;
-import org.zalando.baigan.service.ConfigurationRepository;
+import org.zalando.baigan.repository.ConfigurationRepository;
 
 import static org.hamcrest.CoreMatchers.notNullValue;
 import static org.hamcrest.MatcherAssert.assertThat;
