+++ src/test/java/org/zalando/baigan/proxy/handler/TestConditionsProcessor.java
@@ -1,12 +1,12 @@
 /**
  * Copyright (C) 2015 Zalando SE (http://tech.zalando.com)
- *
+ * <p>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * <p>
  *         http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.zalando.baigan.service;
+package org.zalando.baigan.proxy.handler;
 
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
@@ -82,28 +82,28 @@ public class TestConditionsProcessor {
 
     private Configuration<String> createConditionsForInAppdomain(
             final Set<String> appdomain) {
-        final Condition<String> conditionForAppdomains = new Condition<String>(
+        final Condition<String> conditionForAppdomains = new Condition<>(
                 APPDOMAIN, new In(appdomain), DHL);
 
         final Set<Condition<String>> conditions = ImmutableSet
                 .of(conditionForAppdomains);
 
-        return new Configuration<String>("express.service.provider",
+        return new Configuration<>("express.service.provider",
                 "Express service provider", conditions, NONE);
     }
 
     private Configuration<Boolean> createMultiConditionsForInAppdomainOrCustomerNumberEndsWith(
             final Set<String> appdomain, final Set<String> customerNumbers) {
-        final Condition<Boolean> conditionForAppdomains = new Condition<Boolean>(
+        final Condition<Boolean> conditionForAppdomains = new Condition<>(
                 APPDOMAIN, new In(appdomain), true);
 
-        final Condition<Boolean> conditionForCustomerNumber = new Condition<Boolean>(
+        final Condition<Boolean> conditionForCustomerNumber = new Condition<>(
                 CUSTOMER_NUMBER, new EndsWith(customerNumbers), true);
 
         final Set<Condition<Boolean>> conditions = ImmutableSet
                 .of(conditionForAppdomains, conditionForCustomerNumber);
 
-        return new Configuration<Boolean>("express.feature.toggle",
+        return new Configuration<>("express.feature.toggle",
                 "Feature toggle", conditions, Boolean.FALSE);
     }
 }
