+++ src/test/java/org/zalando/baigan/repository/FileSystemConfigurationRepositoryTest.java
@@ -1,4 +1,4 @@
-package org.zalando.baigan.service;
+package org.zalando.baigan.repository;
 
 import org.junit.jupiter.api.Test;
